\name{BSSVMOU-package}
\alias{BSSVMOU-package}
\alias{BSSVMOU}
\docType{package}
\title{
  Bayesian Semi-parametric Stochastic Velocity Model with Ornstein-Uhlenbeck process prior (B-SSVM-OU)
}
\description{
  B-SSVM-OU can be applied to model growth dynamics (jointly trajectory, velocity, acceleration) in longitudinal data. 
  Furthermore, how B-SSVM-OU can allow us to examine how acceleration or stable velocity is associated with observed covariates.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Jonggyu Baek

Maintainer: Jonggyu Baek <jongguri@umich.edu>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
